refactor: Remove circular dependencies and redundant code
