+++ src/internal/observable/concat.ts
@@ -1,11 +1,13 @@
 import { Observable } from '../Observable';
 import { ObservableInputTuple, SchedulerLike } from '../types';
 import { concatAll } from '../operators/concatAll';
-import { internalFromArray } from './fromArray';
 import { popScheduler } from '../util/args';
+import { from } from './from';
 
 export function concat<T extends readonly unknown[]>(...inputs: [...ObservableInputTuple<T>]): Observable<T[number]>;
-export function concat<T extends readonly unknown[]>(...inputsAndScheduler: [...ObservableInputTuple<T>, SchedulerLike]): Observable<T[number]>;
+export function concat<T extends readonly unknown[]>(
+  ...inputsAndScheduler: [...ObservableInputTuple<T>, SchedulerLike]
+): Observable<T[number]>;
 
 /**
  * Creates an output Observable which sequentially emits all values from the first given
@@ -113,5 +115,5 @@ export function concat<T extends readonly unknown[]>(...inputsAndScheduler: [...
  * Observable subscription on.
  */
 export function concat(...args: any[]): Observable<unknown> {
-  return concatAll()(internalFromArray(args, popScheduler(args)));
+  return concatAll()(from(args, popScheduler(args)));
 }
