+++ src/internal/operators/concat.ts
@@ -1,8 +1,8 @@
 import { ObservableInputTuple, OperatorFunction, SchedulerLike } from '../types';
 import { operate } from '../util/lift';
 import { concatAll } from './concatAll';
-import { internalFromArray } from '../observable/fromArray';
 import { popScheduler } from '../util/args';
+import { from } from '../observable/from';
 
 /** @deprecated Replaced with {@link concatWith}. Will be removed in v8. */
 export function concat<T, A extends readonly unknown[]>(...sources: [...ObservableInputTuple<A>]): OperatorFunction<T, T | A[number]>;
@@ -17,6 +17,6 @@ export function concat<T, A extends readonly unknown[]>(
 export function concat<T, R>(...args: any[]): OperatorFunction<T, R> {
   const scheduler = popScheduler(args);
   return operate((source, subscriber) => {
-    concatAll()(internalFromArray([source, ...args], scheduler)).subscribe(subscriber as any);
+    concatAll()(from([source, ...args], scheduler)).subscribe(subscriber);
   });
 }
