+++ src/internal/operators/exhaustMap.ts
@@ -2,7 +2,7 @@ import { Observable } from '../Observable';
 import { Subscriber } from '../Subscriber';
 import { ObservableInput, OperatorFunction, ObservedValueOf } from '../types';
 import { map } from './map';
-import { innerFrom } from '../observable/from';
+import { innerFrom } from '../observable/innerFrom';
 import { operate } from '../util/lift';
 import { OperatorSubscriber } from './OperatorSubscriber';
