+++ src/internal/operators/skipUntil.ts
@@ -2,7 +2,7 @@ import { Observable } from '../Observable';
 import { MonoTypeOperatorFunction } from '../types';
 import { operate } from '../util/lift';
 import { OperatorSubscriber } from './OperatorSubscriber';
-import { innerFrom } from '../observable/from';
+import { innerFrom } from '../observable/innerFrom';
 import { noop } from '../util/noop';
 
 /**
