+++ src/internal/operators/throttle.ts
@@ -3,7 +3,7 @@ import { Subscription } from '../Subscription';
 import { MonoTypeOperatorFunction, ObservableInput } from '../types';
 import { operate } from '../util/lift';
 import { OperatorSubscriber } from './OperatorSubscriber';
-import { innerFrom } from '../observable/from';
+import { innerFrom } from '../observable/innerFrom';
 
 export interface ThrottleConfig {
   leading?: boolean;
